% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{UI Setup for Shiny App}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Internal parameter for \code{{shiny}}. Do not remove.}
}
\description{
This function sets up the user interface for the Shiny app, arranging inputs
and outputs in a sidebar layout.
}
