% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_upload.R
\name{mod_file_upload_server}
\alias{mod_file_upload_server}
\title{Server Logic for File Upload}
\usage{
mod_file_upload_server(id, available_series)
}
\arguments{
\item{id}{A namespace identifier for the server logic}

\item{available_series}{Reactive variable to store the available series}
}
\description{
Handles the server-side logic for file uploads, reading and processing the uploaded files.
}
