% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot.R
\name{mod_plot_ui}
\alias{mod_plot_ui}
\title{Plotting Module UI}
\usage{
mod_plot_ui(id)
}
\arguments{
\item{id}{A namespace identifier for the plot elements}
}
\description{
Constructs UI elements for plotting time series data.
}
