% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_time_series.R
\name{mod_time_series_server}
\alias{mod_time_series_server}
\title{Server Logic for Time Series Generation}
\usage{
mod_time_series_server(id, available_series, series_counters)
}
\arguments{
\item{id}{A namespace identifier for the generation logic}

\item{available_series}{Reactive variable to store the generated series}

\item{series_counters}{Reactive value tracking the number of each type of series}
}
\description{
Handles the generation of time series based on user inputs.
}
