% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_methods.R
\name{plot.TimeSeries}
\alias{plot.TimeSeries}
\title{Plot method for TimeSeries objects}
\usage{
\method{plot}{TimeSeries}(x, ...)
}
\arguments{
\item{x}{An object of class 'TimeSeries'.}

\item{...}{Additional arguments for plotting.}
}
\description{
Visualizes the time series data.
}
