% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_methods.R
\name{print.TimeSeries}
\alias{print.TimeSeries}
\title{Print method for TimeSeries objects}
\usage{
\method{print}{TimeSeries}(x, ...)
}
\arguments{
\item{x}{An object of class 'TimeSeries'.}

\item{...}{Additional arguments for printing.}
}
\description{
Custom print method for TimeSeries objects.
}
