% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_methods.R
\name{summary.TimeSeries}
\alias{summary.TimeSeries}
\title{Summary method for TimeSeries objects}
\usage{
\method{summary}{TimeSeries}(object, ...)
}
\arguments{
\item{object}{An object of class 'TimeSeries'.}

\item{...}{Additional arguments for summary.}
}
\value{
A summary of the time series.
}
\description{
Provides a statistical summary of the time series data.
}
